feat, chore: restructured directory, raw commands
- Broke out commands into functions in commands package
- Moved ui package into cmd folder and renamed tui
- Moved config package into cmd folder
- Moved portal package into internal
- Implemented raw commands for send and receive
